:recycle: [repositories] Move `providerstore` fixture to `tests.fixtures`
