+++ tests/fixtures/repositories/domain/providers.py
@@ -1,12 +1,16 @@
 """Fixtures for cutty.repositories.domain.providers."""
+import pathlib
 from collections.abc import Callable
 from typing import Any
 from typing import Optional
+
+import pytest
 
 from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path
 from cutty.repositories.domain.locations import Location
 from cutty.repositories.domain.providers import Provider
+from cutty.repositories.domain.providers import ProviderStore
 from cutty.repositories.domain.repository import Repository
 from cutty.repositories.domain.revisions import Revision
 
@@ -50,3 +54,11 @@ def dictprovider(mapping: Optional[dict[str, Any]] = None) -> Provider:
         return Repository(location.name, path, revision)
 
     return _
+
+
+@pytest.fixture
+def providerstore(tmp_path: pathlib.Path) -> ProviderStore:
+    """Fixture for a simple provider store."""
+    path = tmp_path / "store"
+    path.mkdir()
+    return lambda providername: lambda url: path
